make id uint64_t (was uint32_t in most places)
id isn't that big an overhead inside libdablooms, and this allows
more flexibility with how id is chosen and used by the user

counting_bloom_init() does not use id or count argument, remove them

+++ godablooms/dablooms.go
@@ -50,12 +50,12 @@ func (sb *ScalingBloom) Check(key []byte) bool {
 	return C.scaling_bloom_check(sb.cfilter, cKey, C.size_t(len(key))) == 1
 }
 
-func (sb *ScalingBloom) Add(key []byte, id C.uint32_t) bool {
+func (sb *ScalingBloom) Add(key []byte, id C.uint64_t) bool {
 	cKey := (*C.char)(unsafe.Pointer(&key[0]))
 	return C.scaling_bloom_add(sb.cfilter, cKey, C.size_t(len(key)), id) == 1
 }
 
-func (sb *ScalingBloom) Remove(key []byte, id C.uint32_t) bool {
+func (sb *ScalingBloom) Remove(key []byte, id C.uint64_t) bool {
 	cKey := (*C.char)(unsafe.Pointer(&key[0]))
 	return C.scaling_bloom_remove(sb.cfilter, cKey, C.size_t(len(key)), id) == 1
 }
